+++ grace-provider-dept-8001/src/main/java/com/lgz/grace/provider/dept/service/DeptService.java
@@ -11,7 +11,7 @@ public interface DeptService {
 
     public boolean addDept(Dept dept);
 
-    public Dept findById(Long id);
+    public Dept findById(Integer id);
 
     public List<Dept> findAll();
 }
