+++ grace-provider-dept-8001/src/main/java/com/lgz/grace/provider/dept/service/impl/DeptServiceImpl.java
@@ -23,7 +23,7 @@ public class DeptServiceImpl implements DeptService{
     }
 
     @Override
-    public Dept findById(Long deptId) {
+    public Dept findById(Integer deptId) {
         return deptDao.findById(deptId);
     }
